v1 works! many bugs fixed and leaderboard added

+++ the_house/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, BigInteger, ForeignKey, String, UniqueConstraint
+from sqlalchemy import Column, Integer, BigInteger, ForeignKey, String, UniqueConstraint, Float
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import relationship
 
@@ -28,7 +28,7 @@ class Guild(Base):
 class Bank(Base):
     __tablename__ = 'banks'
     id = Column(Integer, primary_key=True)
-    coins = Column(Integer, default=50)
+    coins = Column(Float, default=50)
     user_id = Column(Integer, ForeignKey('users.id'), nullable=False)
     guild_id = Column(Integer, ForeignKey('guilds.id'), nullable=False)
